requirements.txt
africastalking==1.2.7
asgiref==3.8.1
Authlib==1.3.1
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
cryptography==42.0.8
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-crispy-forms==2.2
django-oidc-provider==0.8.2
docker==7.1.0
filelock==3.15.4
future==1.0.0
gunicorn==22.0.0
idna==3.7
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
pyjwkest==1.4.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-decouple==3.8
python-dotenv==0.21.1
python3-openid==3.2.0
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
schema==0.7.7
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.2
